Cleaned up some -Wconversion warnings.
git-svn-id: https://svn.r-project.org/R/trunk@55743 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/sprintf.c
@@ -275,7 +275,7 @@ SEXP attribute_hidden do_sprintf(SEXP call, SEXP op, SEXP args, SEXP env)
                         _this = a[nthis];
                         if (has_star)
                         {
-                            int nf;
+                            size_t nf;
                             char *p, *q = fmt2;
                             for (p = fmt; *p; p++)
                                 if (*p == '*')
@@ -352,7 +352,7 @@ SEXP attribute_hidden do_sprintf(SEXP call, SEXP op, SEXP args, SEXP env)
                                 if (TYPEOF(_this) != STRSXP)
                                 {
                                     /* as.character method might call sprintf() */
-                                    int nc = strlen(outputString);
+                                    size_t nc = strlen(outputString);
                                     char *z = Calloc(nc + 1, char);
                                     strcpy(z, outputString);
                                     PROTECT(tmp = lang2(install("as.character"), _this));
@@ -470,7 +470,7 @@ SEXP attribute_hidden do_sprintf(SEXP call, SEXP op, SEXP args, SEXP env)
             else
             {                                         /* not '%' : handle string part */
                 char *ch = Rf_strchr(curFormat, '%'); /* MBCS-aware version used */
-                chunk = (ch) ? ch - curFormat : strlen(curFormat);
+                chunk = (ch) ? (size_t)(ch - curFormat) : strlen(curFormat);
                 strncpy(bit, curFormat, chunk);
                 bit[chunk] = '\0';
             }
